Information Page Update
Today I started the Information page.
I coded:
The grid system for the different sections.
Copied over the nav bar and added in Jaime's Logo.
Added in all the content needed within the boxes.
Formatted all text to be as is in the prototype (Colour, font, size, weight, decoration etc).
Created a list for the right hand side information, made it so on hover it underlines, bold and changes to green
All that needs to be done is adding the image behind the header text.