Work 21/05
Today in class we did group work!

Shiv and Jaime worked on the home index page.

- Grid established
- search boxes defined
- all colour, font and styling brought over from figma with a few changes made to be codeable
- text classes defined
- grid classes defined
- lots and lots of troubleshooting lol

Our index page is basically finished!
We just need to clean up the dividers but they proved difficult to get center-aligned within the nav due to padding/margins.

Darcey worked on the draft for our presentation and getting a good understanding of Jaime's assignment in terms of designing for older people and the design decisions on the site. She also uploaded all our images for the category page.

Sophia worked on the info.html page and basically got it all up and running!