Categories page
Coded most of the elements on the categories page. Using grid-template areas was very useful as I was able to adjust where each of the elements were on the page.

One problem I had was fitting everything above the fold, originally everything was above the fold, however after adding the nav bar, some of the section elements were seen by scrolling.

The boxes are slightly wider on this website compared to the design, this was because the images aspect ratio was not 1:1 and instead it was 1:55 width / height. This meant the boxes needed to follow this ratio as well.

Overall, I'm quite happy with how this page is coming along.