Search bar & Filter system
Getting the search bar to function. The search bar took me a few hours to get working because I was having problems accessing all of the list elements through a class which was only returning one result. Because of this, I made three classes for three categories to filter out the words. Still having problems moving in-block elements up on the page when other elements have disappeared.