Coding block
In this coding session I renamed the 'interior architecture' heading on the careers page as this was a duplicate of another category. In its place, I put "product designer", and also replaced the image underneath this heading. Additionally, I also edited some of the images I orginally uploaded. I did this because I wanted these images to relatively coordinate with  the websites colour palette, as well as stand out to the user.