Career-info
- wrote and styled list items in right grid area

- added customer bullet item for list

- resized header image on info.html

still having a really hard time with the grid items for cat.html, none of us can figure out the resizing or the js at this stage which is frustrating because it would be so cool to have the filter work! but having the content hidden gets the intent across and will give us a good talking point in our presentation.