Footer & Career Header
I developed the Footer. Added in 3 rows for the footer, with a paragraph, 4 icons in the middle, and then another paragraph. Added in the fontawesome link to the head.
Had some trouble with alignment, for some reason it wont align properly even when we set margin to 0 and padding to 0. I think it's to do with the icons varying in size maybe?
I also added in the header image on the careers page.